If a bot sends a message it will ignore this for the random message respond fucntion. Changed line 142 of MessageEventListener from selfuser to isbot.
Made a simple solution to the random message respond also counting the botcommands. It makes a char array from the message contents and then looks if the message starts with ? or !. if it does it ignores it. It also ignores pokecord commands. If more custom botcommands need to be excluded it would be wise to convert these two if statements to a fucntion with a list of all the excluded start chars. Thanks @bloxxyy for suggesting to use a char array